+++ CHANGELOG.md
@@ -2,6 +2,7 @@ You're really going to want to read this.
 
 ## Unreleased
 
+* Remove \Request::$current and \Request::current() - HMVC no longer in use
 * Remove all support for external routes - these are no longer relevant now we don't support
   external requests.
 * Extract all request execution from the \Request class - instead, use Request_Executor to execute